refactor color for build move all shared methods to helper add missing comments, fix linter warnings
