+++ commands/token.go
@@ -20,14 +20,9 @@ func tokenCmd(cmd *cli.Cmd) cli.ExitValue {
 		return cli.Failure
 	}
 	env := cmd.Env.(config.TravisCommandConfig)
-	if len(env.Token) > 0 {
-		cmd.Stdout.Print("Your access token for ")
-		cmd.Stdout.Cprint("yellow", env.Endpoint)
-		cmd.Stdout.Print(" is ")
-		cmd.Stdout.Cprintln("boldgreen", env.Token)
-		return cli.Success
-	} else {
-		// cmd.Stderr.Println("Not logged in for " + env.Endpoint + ", please run travis login.")
-		return cli.Failure
-	}
+	cmd.Stdout.Print("Your access token for ")
+	cmd.Stdout.Cprint("yellow", env.Endpoint)
+	cmd.Stdout.Print(" is ")
+	cmd.Stdout.Cprintln("boldgreen", env.Token)
+	return cli.Success
 }
